requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
tortoise-orm[asyncpg]==0.20.0